test(hooks): consolidate AfterAgent tests into hooks-system.test.ts
Moves AfterAgent integration tests into the main hooks-system suite. Converts hook scripts from Node.js to Bash using jq for consistency with other tests in the file. Simplifies the interactive test case to improve stability.
